refact(SlimeBase.cs) : Slime의 확장성으로 위해 SlimeBase로 책임을 분배
