+++ Assets/Scripts/Creature/Minion/Slime/SlimeBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7742cbe1cd2b341b18fc51a9ebaff7aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
